Moving remaining DFA Actions to Integration
  * Removed pre-bolonha module
  * Issue: FIST-225
