+++ fenixedu-ist-integration/src/main/java/pt/ist/fenixedu/integration/service/services/dfa/CreateCandidacy.java
@@ -16,7 +16,7 @@
  * You should have received a copy of the GNU Lesser General Public License
  * along with FenixEdu IST Pre Bolonha.  If not, see <http://www.gnu.org/licenses/>.
  */
-package org.fenixedu.academic.service;
+package pt.ist.fenixedu.integration.service.services.dfa;
 
 import java.util.Locale;
