+++ fenixedu-ist-integration/src/main/java/pt/ist/fenixedu/integration/service/services/dfa/RegisterCandidate.java
@@ -16,7 +16,7 @@
  * You should have received a copy of the GNU Lesser General Public License
  * along with FenixEdu IST Pre Bolonha.  If not, see <http://www.gnu.org/licenses/>.
  */
-package org.fenixedu.academic.service.services.masterDegree.administrativeOffice.candidate;
+package pt.ist.fenixedu.integration.service.services.dfa;
 
 import static org.fenixedu.academic.predicate.AccessControl.check;
