+++ fenixedu-ist-integration/src/main/webapp/masterDegreeAdministrativeOffice/candidacy/candidacyRegistration.jsp
