+++ fenixedu-ist-integration/src/main/webapp/masterDegreeAdministrativeOffice/candidacy/candidacyRegistrationSuccess.jsp
