+++ fenixedu-ist-integration/src/main/webapp/masterDegreeAdministrativeOffice/candidacy/confirmCandidaciesForSelection.jsp
