+++ fenixedu-ist-integration/src/main/webapp/masterDegreeAdministrativeOffice/candidacy/showCandidaciesForSelectionSuccess.jsp
