+++ fenixedu-ist-integration/src/main/webapp/masterDegreeAdministrativeOffice/candidacy/showCandidacyValidateData.jsp
