--- fenixedu-ist-pre-bolonha/pom.xml
@@ -1,27 +1,0 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-
-    <parent>
-        <groupId>pt.ist</groupId>
-        <artifactId>fenixedu-ist</artifactId>
-        <version>7.0.0-SNAPSHOT</version>
-    </parent>
-
-    <artifactId>fenixedu-ist-pre-bolonha</artifactId>
-
-    <name>FenixEdu IST Pre Bolonha</name>
-
-    <dependencies>
-        <dependency>
-            <groupId>org.fenixedu</groupId>
-            <artifactId>fenixedu-academic</artifactId>
-        </dependency>
-    </dependencies>
-
-    <repositories>
-        <repository>
-            <id>fenixedu-maven-repository</id>
-            <url>https://repo.fenixedu.org/fenixedu-maven-repository</url>
-        </repository>
-    </repositories>
-</project>
