+++ pom.xml
@@ -125,7 +125,6 @@
         <module>fenixedu-ist-integration</module>
         <module>fenixedu-ist-libraryspace</module>
         <module>fenixedu-ist-parking</module>
-        <module>fenixedu-ist-pre-bolonha</module>
         <module>fenixedu-ist-quc</module>
         <module>fenixedu-ist-reports</module>
         <module>fenixedu-ist-teacher-evaluation</module>
